Move open_raster in helper.py
